backend/requirements.txt
# Core
fastapi==0.109.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlmodel==0.0.14
neo4j==5.16.0
redis==5.0.1

# Storage
minio==7.2.3

# CLI
click==8.1.7
rich==13.7.0

# API Clients
requests==2.31.0

# AI
openai==1.10.0
anthropic==0.8.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2